Where do I start..
1. Updated Login page
2. Added SignUp page that uses the create user post api call to create a new user. ***NEEDS FIXING*** Currently, the user_id's are getting mixed up. I created a new user called "test123" and now all of the journals associated with "seth" belong to test123. When I login as Seth, there are no journals.
3. "Entry" page didn't update the markdown content when you click on edit. I tried updating this and found a way to get it to immediately show the changes. There is still an issue where it won't update the context, so if you go to a different journal entry it will show the text of the previous entry you were at.

+++ src/views/journal/index.tsx
@@ -97,81 +97,6 @@ interface MatchParams {
     eid: string
   };
 
-const journalEntryItems: JournalEntryObject[] = [
-    // {
-    //     id: "one",
-    //     journalid: "full",
-    //     name: "Adding Terraform to Your Infrastructure",
-    //     date: "1/4/2021",
-    //     count: 189,
-    //     owner: "Drew Christofferson"
-    // },
-    // {
-    //     id: "two",
-    //     journalid: "google",
-    //     name: "Leadership Skills in Tech",
-    //     date: "2/19/2021",
-    //     count: 1109,
-    //     owner: "Drew Christofferson"
-    // },
-    // {
-    //     id: "three",
-    //     journalid: "full",
-    //     name: "Why Docker Make Local Development So Easy",
-    //     date: "1/4/2021",
-    //     count: 710,
-    //     owner: "Drew Christofferson"
-    // },
-    // {
-    //     id: "four",
-    //     journalid: "full",
-    //     name: "Git Commands to Remember",
-    //     date: "1/4/2021",
-    //     count: 189,
-    //     owner: "Drew Christofferson"
-    // },
-    // {
-    //     id: "five",
-    //     journalid: "personal",
-    //     name: "My Company Watchlist 2021",
-    //     date: "2/19/2021",
-    //     count: 1109,
-    //     owner: "Drew Christofferson"
-    // },
-    // {
-    //     id: "six",
-    //     journalid: "full",
-    //     name: "On Spring Boot Security",
-    //     date: "1/4/2021",
-    //     count: 710,
-    //     owner: "Drew Christofferson"
-    // },
-    // {
-    //     id: "seven",
-    //     journalid: "google",
-    //     name: "Kubernetes vs. ECS",
-    //     date: "1/4/2021",
-    //     count: 189,
-    //     owner: "Drew Christofferson"
-    // },
-    // {
-    //     id: "eight",
-    //     journalid: "google",
-    //     name: "Tips for Managing a Team",
-    //     date: "2/19/2021",
-    //     count: 1109,
-    //     owner: "Drew Christofferson"
-    // },
-    // {
-    //     id: "nine",
-    //     journalid: "personal",
-    //     name: "Why Keep a Dev Journal?",
-    //     date: "1/4/2021",
-    //     count: 710,
-    //     owner: "Drew Christofferson"
-    // },
-
-]
 
 function Journal () {
     let history = useHistory();
@@ -194,9 +119,13 @@ function Journal () {
     }
 
     const handleRecordDelete = async (id: string) => {
-        await axios.delete(`${context.API_BASE_URL}/api/v1/record/${id}`, config)
-        .then((res) => getRecords());
-        console.log(context.records)
+        if (window.confirm("Are you sure you want to delete this entry? This cannot be undone.")) { //prompt the user and confirm they want to delete the journal entry
+            //if yes, delete the entry
+            await axios.delete(`${context.API_BASE_URL}/api/v1/record/${id}`, config)
+            .then((res) => getRecords());
+        }
+
+        // console.log(context.records)
     };
 
     useEffect(() => {
